tutorial_ui/api/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
google-cloud-storage==2.18.2
google-auth==2.36.0
httpx==0.28.1
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
